problem1/requirements.txt
# Only one external package per assignment rules
psycopg2-binary>=2.9.9
